plugin for grpc callback
