+++ pkg/runtime/const.go
@@ -19,6 +19,5 @@ package runtime
 import "time"
 
 const (
-	Metadata_key_pubsubName = "pubsubName"
-	dialTimeout             = time.Second * 30
+	dialTimeout = time.Second * 30
 )
