+++ pkg/runtime/runtime.go
@@ -18,7 +18,6 @@ package runtime
 
 import (
 	"context"
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"strings"
@@ -29,13 +28,9 @@ import (
 
 	"mosn.io/layotto/components/file"
 
-	"github.com/dapr/components-contrib/contenttype"
 	"github.com/dapr/components-contrib/pubsub"
 	"github.com/dapr/components-contrib/state"
-	jsoniter "github.com/json-iterator/go"
 	rawGRPC "google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"mosn.io/layotto/components/configstores"
 	"mosn.io/layotto/components/hello"
 	"mosn.io/layotto/components/lock"
@@ -50,7 +45,6 @@ import (
 	runtime_pubsub "mosn.io/layotto/pkg/runtime/pubsub"
 	runtime_sequencer "mosn.io/layotto/pkg/runtime/sequencer"
 	runtime_state "mosn.io/layotto/pkg/runtime/state"
-	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 	mgrpc "mosn.io/mosn/pkg/filter/network/grpc"
 	"mosn.io/pkg/log"
 )
@@ -71,29 +65,19 @@ type MosnRuntime struct {
 	sequencerRegistry   runtime_sequencer.Registry
 	bindingsRegistry    mbindings.Registry
 	// component pool
-	hellos            map[string]hello.HelloService
-	configStores      map[string]configstores.Store
-	rpcs              map[string]rpc.Invoker
-	pubSubs           map[string]pubsub.PubSub
-	topicPerComponent map[string]TopicSubscriptions
-	states            map[string]state.Store
-	files             map[string]file.File
-	locks             map[string]lock.LockStore
-	sequencers        map[string]sequencer.Store
-	outputBindings    map[string]bindings.OutputBinding
+	hellos         map[string]hello.HelloService
+	configStores   map[string]configstores.Store
+	rpcs           map[string]rpc.Invoker
+	pubSubs        map[string]pubsub.PubSub
+	states         map[string]state.Store
+	files          map[string]file.File
+	locks          map[string]lock.LockStore
+	sequencers     map[string]sequencer.Store
+	outputBindings map[string]bindings.OutputBinding
 	// app callback
 	AppCallbackConn *rawGRPC.ClientConn
 	// extends
 	errInt ErrInterceptor
-	json   jsoniter.API
-}
-
-type Details struct {
-	metadata map[string]string
-}
-
-type TopicSubscriptions struct {
-	topic2Details map[string]Details
 }
 
 func NewMosnRuntime(runtimeConfig *MosnRuntimeConfig) *MosnRuntime {
@@ -119,7 +103,6 @@ func NewMosnRuntime(runtimeConfig *MosnRuntimeConfig) *MosnRuntime {
 		locks:               make(map[string]lock.LockStore),
 		sequencers:          make(map[string]sequencer.Store),
 		outputBindings:      make(map[string]bindings.OutputBinding),
-		json:                jsoniter.ConfigFastest,
 	}
 }
 
@@ -186,6 +169,10 @@ func (m *MosnRuntime) Run(opts ...Option) (mgrpc.RegisteredServer, error) {
 			m.sequencers,
 			m.sendToOutputBinding,
 		)
+		// init the GrpcAPI
+		if err := api.Init(m.AppCallbackConn); err != nil {
+			return nil, err
+		}
 		apis = append(apis, api)
 	}
 	// put them into grpc options
@@ -314,27 +301,29 @@ func (m *MosnRuntime) initRpcs(rpcs ...*rpc.Factory) error {
 func (m *MosnRuntime) initPubSubs(factorys ...*runtime_pubsub.Factory) error {
 	// 1. init components
 	log.DefaultLogger.Infof("[runtime] start initializing pubsub components")
-	// register all config store services implementation
+	// register all components implementation
 	m.pubSubRegistry.Register(factorys...)
 	for name, config := range m.runtimeConfig.PubSubManagement {
+		// create component
 		comp, err := m.pubSubRegistry.Create(name)
 		if err != nil {
 			m.errInt(err, "create pubsub component %s failed", name)
 			return err
 		}
+		// check config
 		consumerID := strings.TrimSpace(config.Metadata["consumerID"])
 		if consumerID == "" {
 			config.Metadata["consumerID"] = m.runtimeConfig.AppManagement.AppId
 		}
-
+		// init this component with the config
 		if err := comp.Init(pubsub.Metadata{Properties: config.Metadata}); err != nil {
 			m.errInt(err, "init pubsub component %s failed", name)
 			return err
 		}
+		// register this component
 		m.pubSubs[name] = comp
 	}
-	// 2. start subscribing
-	return m.startSubscribing()
+	return nil
 }
 
 func (m *MosnRuntime) initStates(factorys ...*runtime_state.Factory) error {
@@ -447,184 +436,6 @@ func (m *MosnRuntime) initSequencers(factorys ...*runtime_sequencer.Factory) err
 		m.sequencers[name] = comp
 	}
 	return nil
-}
-
-func (m *MosnRuntime) startSubscribing() error {
-	// 1. check if there is no need to do it
-	if len(m.pubSubs) == 0 {
-		return nil
-	}
-	topicRoutes, err := m.getInterestedTopics()
-	if err != nil {
-		return err
-	}
-	if len(topicRoutes) == 0 {
-		//	no need
-		return nil
-	}
-	// 2. loop subscribe
-	for name, pubsub := range m.pubSubs {
-		if err := m.beginPubSub(name, pubsub, topicRoutes); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (m *MosnRuntime) beginPubSub(pubsubName string, ps pubsub.PubSub, topicRoutes map[string]TopicSubscriptions) error {
-	// 1. call app to find topic topic2Details.
-	v, ok := topicRoutes[pubsubName]
-	if !ok {
-		return nil
-	}
-	// 2. loop subscribing every <topic, route>
-	for topic, route := range v.topic2Details {
-		// TODO limit topic scope
-		log.DefaultLogger.Debugf("[runtime][beginPubSub]subscribing to topic=%s on pubsub=%s", topic, pubsubName)
-		// ask component to subscribe
-		if err := ps.Subscribe(pubsub.SubscribeRequest{
-			Topic:    topic,
-			Metadata: route.metadata,
-		}, func(ctx context.Context, msg *pubsub.NewMessage) error {
-			if msg.Metadata == nil {
-				msg.Metadata = make(map[string]string, 1)
-			}
-			msg.Metadata[Metadata_key_pubsubName] = pubsubName
-			return m.publishMessageGRPC(ctx, msg)
-		}); err != nil {
-			log.DefaultLogger.Warnf("[runtime][beginPubSub]failed to subscribe to topic %s: %s", topic, err)
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (m *MosnRuntime) getInterestedTopics() (map[string]TopicSubscriptions, error) {
-	// 1. check
-	if m.topicPerComponent != nil {
-		return m.topicPerComponent, nil
-	}
-	if m.AppCallbackConn == nil {
-		return make(map[string]TopicSubscriptions), nil
-	}
-	comp2Topic := make(map[string]TopicSubscriptions)
-	var subscriptions []*runtimev1pb.TopicSubscription
-
-	// 2. handle app subscriptions
-	client := runtimev1pb.NewAppCallbackClient(m.AppCallbackConn)
-	subscriptions = runtime_pubsub.ListTopicSubscriptions(client, log.DefaultLogger)
-	// TODO handle declarative subscriptions
-
-	// 3. prepare result
-	for _, s := range subscriptions {
-		if s == nil {
-			continue
-		}
-		if _, ok := comp2Topic[s.PubsubName]; !ok {
-			comp2Topic[s.PubsubName] = TopicSubscriptions{topic2Details: make(map[string]Details)}
-		}
-		comp2Topic[s.PubsubName].topic2Details[s.Topic] = Details{metadata: s.Metadata}
-	}
-
-	// 4. log
-	if len(comp2Topic) > 0 {
-		for pubsubName, v := range comp2Topic {
-			topics := []string{}
-			for topic := range v.topic2Details {
-				topics = append(topics, topic)
-			}
-			log.DefaultLogger.Infof("[runtime][getInterestedTopics]app is subscribed to the following topics: %v through pubsub=%s", topics, pubsubName)
-		}
-	}
-	// 5. cache the result
-	m.topicPerComponent = comp2Topic
-	return comp2Topic, nil
-}
-
-func (m *MosnRuntime) publishMessageGRPC(ctx context.Context, msg *pubsub.NewMessage) error {
-	// 1. Unmarshal to cloudEvent model
-	var cloudEvent map[string]interface{}
-	err := m.json.Unmarshal(msg.Data, &cloudEvent)
-	if err != nil {
-		log.DefaultLogger.Debugf("[runtime]error deserializing cloud events proto: %s", err)
-		return err
-	}
-
-	// 2. Drop msg if the current cloud event has expired
-	if pubsub.HasExpired(cloudEvent) {
-		log.DefaultLogger.Warnf("[runtime]dropping expired pub/sub event %v as of %v", cloudEvent[pubsub.IDField].(string), cloudEvent[pubsub.ExpirationField].(string))
-		return nil
-	}
-
-	// 3. Convert to proto domain struct
-	envelope := &runtimev1pb.TopicEventRequest{
-		Id:              cloudEvent[pubsub.IDField].(string),
-		Source:          cloudEvent[pubsub.SourceField].(string),
-		DataContentType: cloudEvent[pubsub.DataContentTypeField].(string),
-		Type:            cloudEvent[pubsub.TypeField].(string),
-		SpecVersion:     cloudEvent[pubsub.SpecVersionField].(string),
-		Topic:           msg.Topic,
-		PubsubName:      msg.Metadata[Metadata_key_pubsubName],
-	}
-
-	// set data field
-	if data, ok := cloudEvent[pubsub.DataBase64Field]; ok && data != nil {
-		decoded, decodeErr := base64.StdEncoding.DecodeString(data.(string))
-		if decodeErr != nil {
-			log.DefaultLogger.Debugf("unable to base64 decode cloudEvent field data_base64: %s", decodeErr)
-			return err
-		}
-
-		envelope.Data = decoded
-	} else if data, ok := cloudEvent[pubsub.DataField]; ok && data != nil {
-		envelope.Data = nil
-
-		if contenttype.IsStringContentType(envelope.DataContentType) {
-			envelope.Data = []byte(data.(string))
-		} else if contenttype.IsJSONContentType(envelope.DataContentType) {
-			envelope.Data, _ = m.json.Marshal(data)
-		}
-	}
-	// TODO tracing
-
-	// 4. Call appcallback
-	clientV1 := runtimev1pb.NewAppCallbackClient(m.AppCallbackConn)
-	res, err := clientV1.OnTopicEvent(ctx, envelope)
-
-	// 5. Check result
-	return retryStrategy(err, res, cloudEvent)
-}
-
-// retryStrategy returns error when the message should be redelivered
-func retryStrategy(err error, res *runtimev1pb.TopicEventResponse, cloudEvent map[string]interface{}) error {
-	if err != nil {
-		errStatus, hasErrStatus := status.FromError(err)
-		if hasErrStatus && (errStatus.Code() == codes.Unimplemented) {
-			// DROP
-			log.DefaultLogger.Warnf("[runtime]non-retriable error returned from app while processing pub/sub event %v: %s", cloudEvent[pubsub.IDField].(string), err)
-			return nil
-		}
-
-		err = errors.New(fmt.Sprintf("error returned from app while processing pub/sub event %v: %s", cloudEvent[pubsub.IDField].(string), err))
-		log.DefaultLogger.Debugf("%s", err)
-		// on error from application, return error for redelivery of event
-		return err
-	}
-
-	switch res.GetStatus() {
-	case runtimev1pb.TopicEventResponse_SUCCESS:
-		// on uninitialized status, this is the case it defaults to as an uninitialized status defaults to 0 which is
-		// success from protobuf definition
-		return nil
-	case runtimev1pb.TopicEventResponse_RETRY:
-		return errors.New(fmt.Sprintf("RETRY status returned from app while processing pub/sub event %v", cloudEvent[pubsub.IDField].(string)))
-	case runtimev1pb.TopicEventResponse_DROP:
-		log.DefaultLogger.Warnf("[runtime]DROP status returned from app while processing pub/sub event %v", cloudEvent[pubsub.IDField].(string))
-		return nil
-	}
-	// Consider unknown status field as error and retry
-	return errors.New(fmt.Sprintf("unknown status returned from app while processing pub/sub event %v: %v", cloudEvent[pubsub.IDField].(string), res.GetStatus()))
 }
 
 func (m *MosnRuntime) initAppCallbackConnection() error {
