+++ pkg/runtime/runtime_test.go
@@ -17,8 +17,6 @@
 package runtime
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
 	l8grpc "mosn.io/layotto/pkg/grpc"
 	"net"
@@ -30,11 +28,8 @@ import (
 	"github.com/dapr/components-contrib/pubsub"
 	"github.com/dapr/components-contrib/state"
 	"github.com/golang/mock/gomock"
-	jsoniter "github.com/json-iterator/go"
 	"github.com/stretchr/testify/assert"
 	rawGRPC "google.golang.org/grpc"
-	"google.golang.org/grpc/test/bufconn"
-
 	"mosn.io/pkg/log"
 
 	"mosn.io/layotto/components/configstores"
@@ -48,12 +43,10 @@ import (
 	mock_pubsub "mosn.io/layotto/pkg/mock/components/pubsub"
 	mock_sequencer "mosn.io/layotto/pkg/mock/components/sequencer"
 	mock_state "mosn.io/layotto/pkg/mock/components/state"
-	mock_appcallback "mosn.io/layotto/pkg/mock/runtime/appcallback"
 	mlock "mosn.io/layotto/pkg/runtime/lock"
 	mpubsub "mosn.io/layotto/pkg/runtime/pubsub"
 	msequencer "mosn.io/layotto/pkg/runtime/sequencer"
 	mstate "mosn.io/layotto/pkg/runtime/state"
-	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 )
 
 func TestNewMosnRuntime(t *testing.T) {
@@ -128,37 +121,9 @@ func TestMosnRuntime_initAppCallbackConnection(t *testing.T) {
 
 func TestMosnRuntime_initPubSubs(t *testing.T) {
 	t.Run("normal", func(t *testing.T) {
-		// mock callback response
-		subResp := &runtimev1pb.ListTopicSubscriptionsResponse{
-			Subscriptions: []*runtimev1pb.TopicSubscription{
-				{
-					PubsubName: "mock",
-					Topic:      "layotto",
-					Metadata:   nil,
-				},
-			},
-		}
-		// init grpc server
-		mockAppCallbackServer := mock_appcallback.NewMockAppCallbackServer(gomock.NewController(t))
-		mockAppCallbackServer.EXPECT().ListTopicSubscriptions(gomock.Any(), gomock.Any()).Return(subResp, nil)
-
-		lis := bufconn.Listen(1024 * 1024)
-		s := rawGRPC.NewServer()
-		runtimev1pb.RegisterAppCallbackServer(s, mockAppCallbackServer)
-		go func() {
-			s.Serve(lis)
-		}()
-
-		// init callback client
-		callbackClient, err := rawGRPC.DialContext(context.Background(), "bufnet", rawGRPC.WithInsecure(), rawGRPC.WithContextDialer(func(ctx context.Context, s string) (net.Conn, error) {
-			return lis.Dial()
-		}))
-		assert.Nil(t, err)
-
 		// mock pubsub component
 		mockPubSub := mock_pubsub.NewMockPubSub(gomock.NewController(t))
 		mockPubSub.EXPECT().Init(gomock.Any()).Return(nil)
-		mockPubSub.EXPECT().Subscribe(gomock.Any(), gomock.Any()).Return(nil)
 		f := func() pubsub.PubSub {
 			return mockPubSub
 		}
@@ -174,12 +139,11 @@ func TestMosnRuntime_initPubSubs(t *testing.T) {
 		}
 		// construct MosnRuntime
 		m := NewMosnRuntime(cfg)
-		m.AppCallbackConn = callbackClient
 		m.errInt = func(err error, format string, args ...interface{}) {
 			log.DefaultLogger.Errorf("[runtime] occurs an error: "+err.Error()+", "+format, args...)
 		}
 		// test initPubSubs
-		err = m.initPubSubs(mpubsub.NewFactory("mock", f))
+		err := m.initPubSubs(mpubsub.NewFactory("mock", f))
 		// assert result
 		assert.Nil(t, err)
 	})
@@ -325,58 +289,6 @@ func TestMosnRuntime_initLocks(t *testing.T) {
 			log.DefaultLogger.Errorf("[runtime] occurs an error: "+err.Error()+", "+format, args...)
 		}
 		err := m.initLocks(mlock.NewFactory("mock", f))
-		assert.Nil(t, err)
-	})
-}
-
-func TestMosnRuntime_publishMessageGRPC(t *testing.T) {
-	t.Run("publish success", func(t *testing.T) {
-		subResp := &runtimev1pb.TopicEventResponse{
-			Status: runtimev1pb.TopicEventResponse_SUCCESS,
-		}
-		// init grpc server
-		mockAppCallbackServer := mock_appcallback.NewMockAppCallbackServer(gomock.NewController(t))
-		mockAppCallbackServer.EXPECT().OnTopicEvent(gomock.Any(), gomock.Any()).Return(subResp, nil)
-
-		lis := bufconn.Listen(1024 * 1024)
-		s := rawGRPC.NewServer()
-		runtimev1pb.RegisterAppCallbackServer(s, mockAppCallbackServer)
-		go func() {
-			s.Serve(lis)
-		}()
-
-		// init callback client
-		callbackClient, err := rawGRPC.DialContext(context.Background(), "bufnet", rawGRPC.WithInsecure(), rawGRPC.WithContextDialer(func(ctx context.Context, s string) (net.Conn, error) {
-			return lis.Dial()
-		}))
-		assert.Nil(t, err)
-
-		cloudEvent := map[string]interface{}{
-			pubsub.IDField:              "id",
-			pubsub.SourceField:          "source",
-			pubsub.DataContentTypeField: "content-type",
-			pubsub.TypeField:            "type",
-			pubsub.SpecVersionField:     "v1.0.0",
-			pubsub.DataBase64Field:      "bGF5b3R0bw==",
-		}
-
-		data, err := json.Marshal(cloudEvent)
-		assert.Nil(t, err)
-
-		msg := &pubsub.NewMessage{
-			Data:     data,
-			Topic:    "layotto",
-			Metadata: make(map[string]string),
-		}
-
-		cfg := &MosnRuntimeConfig{}
-		m := NewMosnRuntime(cfg)
-		m.errInt = func(err error, format string, args ...interface{}) {
-			log.DefaultLogger.Errorf("[runtime] occurs an error: "+err.Error()+", "+format, args...)
-		}
-		m.AppCallbackConn = callbackClient
-		m.json = jsoniter.ConfigFastest
-		err = m.publishMessageGRPC(context.Background(), msg)
 		assert.Nil(t, err)
 	})
 }
